move logical query processor to logical folder
