+++ snuba/query/processors/logical/quota_processor.py
@@ -1,6 +1,6 @@
 from snuba.clickhouse.query_dsl.accessors import get_object_ids_in_query_ast
 from snuba.query.logical import Query
-from snuba.query.processors import QueryProcessor
+from snuba.query.processors.logical import LogicalQueryProcessor
 from snuba.query.query_settings import QuerySettings
 from snuba.state import get_config
 from snuba.state.quota import ResourceQuota
@@ -9,7 +9,7 @@ ENABLED_CONFIG = "resource_quota_processor_enabled"
 REFERRER_PROJECT_CONFIG = "referrer_project_thread_quota"
 
 
-class ResourceQuotaProcessor(QueryProcessor):
+class ResourceQuotaProcessor(LogicalQueryProcessor):
     """
     Applies a referrer/project thread quota to the query.
     """
