+++ snuba/query/processors/logical/tags_expander.py
@@ -2,11 +2,11 @@ from dataclasses import replace
 
 from snuba.query.expressions import Column, Expression, FunctionCall
 from snuba.query.logical import Query
-from snuba.query.processors import QueryProcessor
+from snuba.query.processors.logical import LogicalQueryProcessor
 from snuba.query.query_settings import QuerySettings
 
 
-class TagsExpanderProcessor(QueryProcessor):
+class TagsExpanderProcessor(LogicalQueryProcessor):
     """
     Transforms the special syntax we provide to expand the tags column into a call
     to arrayJoin so that, after this query processor, tags_key and tags_value special
